Separated coupons into Ready to Use and Not Yet Actioned as to new Wireframes
The previous logic was no longer suitable and there is now three columns
and the mobile view needs to be progressively loaded. But there is a bug
involving the coupons not remembered their state on resize, e.g.
clicking sent to card on desktop will not say Sent to Card on tablet